Extract page layout
