+++ ui/src/app/cohort-review/cohort-review/cohort-review.component.html
@@ -1,4 +1,4 @@
-<div #wrapper class="cohort-review-wrapper">
+<app-fullpage-app>
   <div *ngIf="loading" class="spinner">Loading...</div>
 
   <!-- Main subject navigation sidebar -->
@@ -133,4 +133,4 @@
       </button>
     </div>
   </clr-modal>
-</div>
+</app-fullpage-app>
